Make it all 64bit safe

+++ bgjslibrary/jni/bgjs/jniext.h
@@ -9,7 +9,7 @@ extern "C" {
 			jobject obj, jlong ctxPtr, jstring path);
 	JNIEXPORT bool JNICALL Java_ag_boersego_bgjs_ClientAndroid_ajaxSuccess(
 			JNIEnv * env, jobject obj, jlong ctxPtr, jstring data,
-			jint responseCode, jint cbPtr, jint thisPtr);
+			jint responseCode, jlong cbPtr, jlong thisPtr);
 	JNIEXPORT void JNICALL Java_ag_boersego_bgjs_ClientAndroid_run(JNIEnv * env,
 			jobject obj, jlong ctxPtr);
 	JNIEXPORT void JNICALL Java_ag_boersego_bgjs_ClientAndroid_timeoutCB(
@@ -17,21 +17,21 @@ extern "C" {
 	JNIEXPORT void JNICALL Java_ag_boersego_bgjs_ClientAndroid_runCBBoolean (JNIEnv * env, jobject obj, jlong ctxPtr, jlong cbPtr, jlong thisPtr, jboolean b);
 
 	// BGJSGLModule
-	JNIEXPORT jint JNICALL Java_ag_boersego_bgjs_ClientAndroid_createGL(JNIEnv * env,
+	JNIEXPORT jlong JNICALL Java_ag_boersego_bgjs_ClientAndroid_createGL(JNIEnv * env,
 			jobject obj, jlong ctxPtr, jobject javaGlView, jfloat pixelRatio, jboolean noClearOnFlip);
 	JNIEXPORT bool JNICALL Java_ag_boersego_bgjs_ClientAndroid_step(JNIEnv * env,
-			jobject obj, jlong ctxPtr, jint jsPtr);
+			jobject obj, jlong ctxPtr, jlong jsPtr);
 	JNIEXPORT int JNICALL Java_ag_boersego_bgjs_ClientAndroid_init(JNIEnv * env,
-            jobject obj, jlong ctxPtr, jint objPtr, jint width, jint height, jstring callbackName);
+            jobject obj, jlong ctxPtr, jlong objPtr, jint width, jint height, jstring callbackName);
 	JNIEXPORT void JNICALL Java_ag_boersego_bgjs_ClientAndroid_setTouchPosition(JNIEnv * env,
-			jobject obj, jlong ctxPtr, jint jsPtr, jint x, jint y);
+			jobject obj, jlong ctxPtr, jlong jsPtr, jint x, jint y);
 	JNIEXPORT void JNICALL Java_ag_boersego_bgjs_ClientAndroid_sendTouchEvent(
-			JNIEnv * env, jobject obj, jlong ctxPtr, jint objPtr, jstring typeStr,
+			JNIEnv * env, jobject obj, jlong ctxPtr, jlong objPtr, jstring typeStr,
 			jfloatArray xArr, jfloatArray yArr, jfloat scale);
 	JNIEXPORT void JNICALL Java_ag_boersego_bgjs_ClientAndroid_close(JNIEnv * env,
-			jobject obj, jlong ctxPtr, jint jsPtr);
+			jobject obj, jlong ctxPtr, jlong jsPtr);
 	JNIEXPORT void JNICALL Java_ag_boersego_bgjs_ClientAndroid_redraw(JNIEnv * env,
-			jobject obj, jlong ctxPtr, jint jsPtr);
+			jobject obj, jlong ctxPtr, jlong jsPtr);
 };
 
 #endif
